Refactor move key in a separated module
