+++ tests/integration.rs
@@ -19,7 +19,7 @@ use hex_literal::hex;
 
 #[test]
 fn should_verify_valid_proof() {
-    let key = AugmentedVerificationKey::default();
+    let key = VerificationKey::default();
     let data = hex!(
         r#"
         283e3f25323d02dabdb94a897dc2697a3b930d8781381ec574af89a201a91d5a
@@ -51,5 +51,5 @@ fn should_verify_valid_proof() {
     let proof = Proof::try_from(&data).unwrap();
     let pubs = hex!("0d69b94acdfaca5bacc248a60b35b925a2374644ce0c1205db68228c8921d9d9").into();
 
-    verify(&key, &proof, &pubs).unwrap();
+    verify(&key.into(), &proof, &pubs).unwrap();
 }
